requirements.txt
streamlit
pandas
presidio-analyzer[transformers]
presidio-analyzer[stanza]
presidio-anonymizer
xlsxwriter
python-dotenv
st-annotated-text
torch
flair
stqdm
streamlit-tags
openpyxl
